build: fix clang-tidy warnings and add core-driver split report
- Add deleted copy and move operations to CeceDriverOrchestrator to satisfy Rule of Five in cppcoreguidelines-special-member-functions.
- Cast multiplier operands to prevent implicit widening in pointer offsets and vector initialization.
- Provide fallback catch block assignments to avoid bugprone-empty-catch.
- Add docs/cece_core_driver_split_report.md outlining physical separation architecture and benefits.

+++ src/main.cpp
@@ -96,7 +96,7 @@ int main(int argc, char* argv[]) {
         if (config["output"] && config["output"]["fields"]) {
             for (const auto& field_node : config["output"]["fields"]) {
                 std::string field_name = field_node.as<std::string>();
-                std::vector<double> field_mem(nx * ny * nz, 0.0);
+                std::vector<double> field_mem(static_cast<std::size_t>(nx) * ny * nz, 0.0);
                 cece_core_set_export_field(cece_data_ptr, field_name.c_str(), static_cast<int>(field_name.length()), field_mem.data(), nx, ny, nz,
                                            &rc);
             }
